Several bugfixes concerning big waveform data (thanks to Simon for providing test data)

+++ plugins/simulator/waveform_viewer/include/waveform_viewer/wave_tree_model.h
@@ -31,7 +31,7 @@ namespace hal {
         Q_OBJECT
         WaveItem* mItem;
         QDir mWorkDir;
-        float mTpos;
+        u64 mTpos;
         int mXpos;
         int mValue;
         bool mAbort;
@@ -40,7 +40,7 @@ namespace hal {
     private Q_SLOTS:
         void handleValueThreadFinished();
     public:
-        WaveValueThread(WaveItem* item, const QString& workdir, float tpos, int xpos, QObject* parent = nullptr);
+        WaveValueThread(WaveItem* item, const QString& workdir, u64 tpos, int xpos, QObject* parent = nullptr);
         void run() override;
         void abort() { mAbort = true; }
         bool wasAborted() const { return mAbort; }
@@ -71,7 +71,7 @@ namespace hal {
         QModelIndexList mDragIndexList;
         DragCommand mDragCommand;
         bool mDragIsGroup;
-        double mCursorTime;
+        u64 mCursorTime;
         int mCursorXpos;
         bool mIgnoreSignals;
         int mReorderRequestWaiting;
@@ -102,7 +102,7 @@ namespace hal {
         void handleGroupAdded(int grpId);
         void handleGroupAboutToBeRemoved(WaveDataGroup* grp);
         void handleGroupUpdated(int grpId);
-        void handleCursorMoved(double tCursor, int xpos);
+        void handleCursorMoved(u64 tCursor, int xpos);
         void forwardBeginResetModel();
         void forwardEndResetModel();
         //void handleValueLoaderFinished();
@@ -147,7 +147,7 @@ namespace hal {
         void addWaves(const QVector<WaveData*>& wds);
         bool onlyRootItemsSelected(const QModelIndexList& selectList) const;
 
-        double cursorTime() const { return mCursorTime; }
+        u64 cursorTime() const { return mCursorTime; }
         int cursorXpos() const { return mCursorXpos; }
 
         QSet<int> waveDataIndexSet() const;
